feat: implement task in 'modules' section
